requirements.txt
streamlit
langchain
langchain-community
langchain-groq
pypdf
python-docx
docx2txt
requests
unstructured
chromadb
sentence-transformers
pandas
plotly
openai
groq
python-dotenv
fastapi
uvicorn
markdown
nltk
flask
tf-keras